refactor to cpp
